requirements.txt
opencv-python
numpy
open3d
torch
torchvision
transformers
scipy
matplotlib
pymavlink
Pillow
pyyaml
